Adding controllers in Backend
As part of the code review I'm simplifying the back end server.js file by using module exports & controllers
