+++ Backend Api/server.js
@@ -2,8 +2,11 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
-const knex = require('knex')
-
+const knex = require('knex');
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image = require('./controllers/imageCount');
 
 const db = knex({
   client: 'pg',
@@ -23,86 +26,14 @@ const app = express();
 app.use(cors())
 app.use(bodyParser.json());
 
-app.post('/signin', (req,res) => {
-	db.select('email', 'hash').from('login')
-	.where('email','=', req.body.email)
-	.then(data => {
-		const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-		console.log(isValid);
-		if (isValid) {
-			return  db.select('*').from('users')
-					.where('email', '=', req.body.email)
-					.then(user => {
-						console.log(user);
-						res.json(user[0])
-					})
-					.catch( err => res.status(400).json('unable to get user'))
-		} else {
-		res.status(400).json('wrong credentials')
-	}
-	})
-	.catch( err => res.status(400).json('wrong credentials'))
-})
-
-
-app.post('/register',(req,res) => {
-	const {email, name, password} = req.body;
-	const hash = bcrypt.hashSync(password);
-	db.transaction(trx => {
-		trx.insert({
-			hash: hash,
-			email: email
-		})
-		.into('login')
-		.returning('email')
-		.then(loginEmail => {
-			return trx('users')
-					.returning('*')
-					.insert({
-						name: name,
-						email: loginEmail[0].email,
-						joined: new Date()
-						})
-					.then(user => {	
-					res.json(user[0]);
-					})
-		})
-		.then(trx.commit)
-		.catch(trx.rollback)
-	})
-	.catch(err => res.status(400).json('unable to register') &&
-	 console.log(err))
-})
-
-
-app.get('/profile/:id', (req, res) => {
-	const { id } = req.params;
-	db.select('*').from('users').where({id})
-	.then(user => {
-		if (user.length) {
-		res.json(user[0])	
-	} else { 
-		res.status(400).json('Not found')
-	}
-	})
-	.catch( err => res.status(400).json('error getting user'))
-})
-
-
-
-app.put('/image', (req,res) => {
-const { id } = req.body;
-db('users').where('id', '=',id)
-.increment('entries',1)
-.returning('entries')
-.then(entries => {
-	res.json(entries[0].entries);
-})
-.catch( err => res.status(400).json('unable to get entries'))
-})
-
-	
-
+app.post('/signin', (req,res) => {signin.handleSignin(req, res, db, bcrypt)}
+	)
+app.post('/register', (req, res) => {register.handleRegister(req, res, db, bcrypt)}
+	)
+app.get('/profile/:id', (req,res) => {profile.handleProfile(req, res, db)}
+	)
+app.put('/image',(req, res) => {image.imageCount(req, res, db)}
+	)
 app.listen(3003, () => {
 	console.log('app is running on port 3003');
 })
